Minor reorganisation in response to review comments
